requirements.txt
asgiref==3.4.1
astroid==2.8.6
Babel==2.9.1
colorama==0.4.4
Django==3.2.9
django-admin-volt==0.0.3
django-ckeditor==6.2.0
django-js-asset==1.2.2
django-phonenumber-field==6.0.0
django-widget-tweaks==1.4.9
gunicorn==20.1.0
isort==5.10.1
lazy-object-proxy==1.6.0
mccabe==0.6.1
phonenumbers==8.12.40
phonenumberslite==8.12.40
Pillow==8.4.0
platformdirs==2.4.0
pylint==2.11.1
pylint-django==2.4.4
pylint-plugin-utils==0.6
pytz==2021.3
sqlparse==0.4.2
toml==0.10.2
wrapt==1.13.3
